refactor: Integrate functions related to Modal
