+++ components/Header/Header.tsx
@@ -2,8 +2,8 @@ import Image from 'next/image';
 import React, { useMemo, useState } from 'react';
 
 import { useDateContext } from '../../contexts/DateContext';
+import { MODAL_NAMES, useModal } from '../../contexts/ModalContext';
 import { useSidebarContext } from '../../contexts/SidebarContext';
-import { useModal } from '../../lib/hooks/useModal';
 import apps_icon from '../../public/images/apps_icon.svg';
 import back_icon from '../../public/images/back_icon.svg';
 import before_icon from '../../public/images/before_icon.svg';
@@ -11,7 +11,6 @@ import calendar_icon from '../../public/images/calendar_icon.svg';
 import menu_icon from '../../public/images/menu_icon.svg';
 import next_icon from '../../public/images/next_icon.svg';
 import search_icon from '../../public/images/search_icon.svg';
-import { MODAL_NAMES } from '../ModalContainer';
 
 import CalendarTypeDropDown from './CalendarTypeDropDown';
 import styles from './Header.module.scss';
