+++ pages/_app.tsx
@@ -2,9 +2,8 @@ import type { AppProps } from 'next/app';
 import React from 'react';
 
 import Header from '../components/Header/Header';
-import ModalContainer from '../components/ModalContainer';
 import DateProvider from '../contexts/DateContext';
-import ModalProvider from '../contexts/ModalContext';
+import ModalProvider, { ModalContainer } from '../contexts/ModalContext';
 import SessionProvider from '../contexts/SessionContext';
 import SidebarProvider from '../contexts/SidebarContext';
 import '../styles/global.css';
